python/requirements.txt
pandas
python-dotenv
sqlalchemy
Flask
requests
pymongo